refactor: move to existing account repository and service
