+++ app/src/main/java/com/example/lunark/util/ClientUtils.java
@@ -1,17 +1,11 @@
 package com.example.lunark.util;
 
 import com.example.lunark.BuildConfig;
-import com.example.lunark.DaggerApplicationComponent;
-import com.example.lunark.clients.LoginService;
-import com.example.lunark.interceptors.JwtInterceptor;
-import com.example.lunark.services.AccountService;
 import com.example.lunark.services.PropertyService;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 
 import java.time.LocalDate;
-import java.util.Date;
-import java.util.Locale;
 import java.util.concurrent.TimeUnit;
 
 import okhttp3.OkHttpClient;
@@ -50,5 +44,4 @@ public class ClientUtils {
             .build();
 
     public static PropertyService propertyService = retrofit.create(PropertyService.class);
-    public static AccountService accountService = retrofit.create(AccountService.class);
 }
